requirements.txt
fastapi[all]
psycopg2
redis
schedule
python-dotenv
pydantic
SQLAlchemy==1.4.0
httpx



